+++ ChangeLog.md
@@ -10,6 +10,7 @@ Copyright (C) 2012, 2019 David M. Syzdek <david@syzdek.net>
   - ldap2csv: adding ability to specify default values (syzdek)
   - ldap2csv: making search filter optional (syzdek)
   - ldap2json: adding utility (syzdek)
+  - ldapinfo: adding utility (syzdek)
   - ldaptree: fixing filter handling when specifing attributes (syzdek)
   - ldaptree: adding comments above graph (syzdek)
   - ldaptree: fixing segfault when displaying linear tree (syzdek)